fix(hex): close settings and marketplace refactor gaps
