+++ src/main/java/me/golemcore/bot/adapter/outbound/storage/LocalWorkspaceFileAdapter.java
@@ -32,7 +32,7 @@ public class LocalWorkspaceFileAdapter implements WorkspaceFilePort {
     }
 
     @Override
-    public Path toRealPath(Path path) throws IOException {
+    public Path resolveRealPath(Path path) throws IOException {
         return path.toRealPath();
     }
