+++ src/main/java/me/golemcore/bot/application/settings/RuntimeSettingsFacade.java
@@ -267,6 +267,7 @@ public class RuntimeSettingsFacade {
     public RuntimeConfig updateMcpCatalogEntry(String name, RuntimeConfig.McpCatalogEntry entry) {
         String normalizedName = validator.normalizeCatalogEntryName(name);
         validator.validateMcpCatalogEntry(entry);
+        entry.setName(normalizedName);
         boolean updated = runtimeConfigService.updateMcpCatalogEntry(normalizedName, entry);
         if (!updated) {
             throw new NoSuchElementException("MCP catalog entry '" + normalizedName + "' not found");
@@ -338,7 +339,8 @@ public class RuntimeSettingsFacade {
             config.setCompaction(compactionConfig);
         }
         runtimeConfigService.updateRuntimeConfig(config);
-        return runtimeConfigService.getRuntimeConfig();
+        RuntimeConfig apiView = runtimeConfigService.getRuntimeConfigForApi();
+        return apiView != null ? apiView : config;
     }
 
     private RuntimeConfig.LlmConfig ensureLlmConfig(RuntimeConfig config) {
