+++ src/test/java/me/golemcore/bot/domain/selfevolving/promotion/PromotionWorkflowStateServiceTest.java
@@ -23,6 +23,7 @@ import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import me.golemcore.bot.adapter.outbound.selfevolving.JsonPromotionWorkflowStateAdapter;
 import me.golemcore.bot.domain.selfevolving.artifact.ArtifactBundleService;
 import me.golemcore.bot.domain.selfevolving.artifact.EvolutionArtifactIdentityService;
 import me.golemcore.bot.domain.selfevolving.candidate.EvolutionCandidateDerivationService;
@@ -77,7 +78,8 @@ class PromotionWorkflowStateServiceTest {
 
         Clock clock = Clock.fixed(Instant.parse("2026-03-31T16:00:00Z"), ZoneOffset.UTC);
         RuntimeConfigService runtimeConfigService = mock(RuntimeConfigService.class);
-        TacticRecordService tacticRecordService = new TacticRecordService(new InMemoryTacticRecordStorePort(), clock, null, null);
+        TacticRecordService tacticRecordService = new TacticRecordService(new InMemoryTacticRecordStorePort(), clock,
+                null, null);
         ArtifactBundleService artifactBundleService = new ArtifactBundleService(
                 new JsonArtifactRepositoryAdapter(storagePort), runtimeConfigService,
                 clock);
@@ -88,7 +90,7 @@ class PromotionWorkflowStateServiceTest {
                 new EvolutionCandidateDerivationService(clock),
                 new EvolutionCandidateTacticMaterializer(clock));
         stateService = new PromotionWorkflowStateService(
-                new PromotionWorkflowStore(new InMemoryPromotionWorkflowStatePort()),
+                new PromotionWorkflowStore(new JsonPromotionWorkflowStateAdapter(storagePort)),
                 evolutionCandidateService,
                 new PromotionDecisionHydrationService());
     }
