+++ src/test/java/me/golemcore/bot/domain/service/SkillMarketplaceServiceTest.java
@@ -14,6 +14,8 @@ import me.golemcore.bot.domain.service.SkillDocumentService;
 import me.golemcore.bot.domain.service.SkillService;
 import me.golemcore.bot.domain.service.WorkspacePathService;
 import me.golemcore.bot.infrastructure.config.BotProperties;
+import me.golemcore.bot.port.outbound.SkillMarketplaceArtifactPort;
+import me.golemcore.bot.port.outbound.SkillMarketplaceCatalogPort;
 import me.golemcore.bot.port.outbound.StoragePort;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
@@ -363,6 +365,69 @@ class SkillMarketplaceLegacySupportTest {
     }
 
     @Test
+    void shouldLoadArtifactContentFromCatalogSourceSnapshotWhenRuntimeSourceChanges(@TempDir Path tempDir)
+            throws Exception {
+        Path firstRepoRoot = createLocalRegistry(tempDir.resolve("first"), """
+                ---
+                name: code-reviewer
+                description: First source skill
+                model_tier: smart
+                ---
+
+                First source body.
+                """);
+        Path secondRepoRoot = createLocalRegistry(tempDir.resolve("second"), """
+                ---
+                name: code-reviewer
+                description: Second source skill
+                model_tier: smart
+                ---
+
+                Second source body.
+                """);
+
+        InMemoryStoragePort storagePort = new InMemoryStoragePort();
+        SkillService skillService = mock(SkillService.class);
+        RuntimeConfigService runtimeConfigService = mock(RuntimeConfigService.class);
+        when(runtimeConfigService.getRuntimeConfig())
+                .thenReturn(RuntimeConfig.builder()
+                        .skills(RuntimeConfig.SkillsConfig.builder()
+                                .marketplaceSourceType("directory")
+                                .marketplaceRepositoryDirectory(firstRepoRoot.toString())
+                                .build())
+                        .build())
+                .thenReturn(RuntimeConfig.builder()
+                        .skills(RuntimeConfig.SkillsConfig.builder()
+                                .marketplaceSourceType("directory")
+                                .marketplaceRepositoryDirectory(secondRepoRoot.toString())
+                                .build())
+                        .build());
+
+        BotProperties properties = botProperties("http://127.0.0.1:1");
+        SkillMarketplaceLegacySupport service = createService(
+                properties,
+                storagePort,
+                skillService,
+                runtimeConfigService);
+        SkillMarketplaceCatalogPort.SkillMarketplaceCatalogData catalogData = service.loadCatalog(
+                me.golemcore.bot.support.TestPorts.settings(properties).skills().marketplace(),
+                Map.of("marketplaceSourceType", "directory",
+                        "marketplaceRepositoryDirectory", firstRepoRoot.toString()),
+                firstRepoRoot.toString(),
+                null,
+                null,
+                null);
+
+        SkillMarketplaceArtifactPort.InstalledArtifactContent artifactContent = service.loadArtifactContent(
+                catalogData.source(),
+                catalogData.artifacts().get("golemcore/code-reviewer"),
+                "golemcore/code-reviewer");
+
+        assertTrue(artifactContent.skillDocuments().getFirst().content().contains("First source body."));
+        assertFalse(artifactContent.skillDocuments().getFirst().content().contains("Second source body."));
+    }
+
+    @Test
     void shouldCleanupPartialInstallWhenSkillWriteFails(@TempDir Path tempDir) throws Exception {
         Path repoRoot = createLocalRegistry(tempDir);
 
@@ -612,7 +677,8 @@ class SkillMarketplaceLegacySupportTest {
 
         BotProperties properties = botProperties("http://127.0.0.1:1");
         properties.getSkills().setMarketplaceEnabled(false);
-        SkillMarketplaceLegacySupport service = createService(properties, storagePort, skillService, runtimeConfigService);
+        SkillMarketplaceLegacySupport service = createService(properties, storagePort, skillService,
+                runtimeConfigService);
 
         IllegalStateException error = assertThrows(IllegalStateException.class,
                 () -> service.install("golemcore/code-reviewer"));
